requirements.txt
requests>=2.32.0
pandas>=2.0.0
tabulate>=0.9.0
numpy>=1.24.0
python-dotenv>=1.0.0
